add: new methods for MetricsFormatter

+++ src/main/java/com/example/metric_api/entitiy/Metrics.java
@@ -34,7 +34,7 @@ public class Metrics {
     private Double processCpuLoad;
 	private Double systemCpuLoad;
 	private Double systemAverageLoad;
-	private Double cpuTemp;
+	private Integer cpuTemp;
 
     // RAM
     private Long memoryUsage;
